scripts/requirements_pipeline.txt
# Dependencias para el Pipeline Unificado de Papers
# AURA360 - Validación, Upload a GCS e Ingesta a Qdrant

# Core
pydantic>=2.6.0
pydantic-settings>=2.2.0
pyyaml>=6.0

# HTTP Client (async)
httpx>=0.25.0

# Cloud Storage - GCS
google-cloud-storage>=3.4.0

# Cloud Storage - S3 (opcional)
# boto3>=1.34.0
# botocore>=1.34.0

# Logging (opcional, ya incluido en Python stdlib)
# loguru>=0.7.0
